Completed the header section
I couldn't comprehend a situation where the second child of a parent container needed to match it's margin with that of the parent containers sibling. It took a lot of my time but i finally realized that i could give margin equal to that of the first child element of the first parent container to the second parent element and then add padding equal to that of the margin required to the next parent container. Anyways i'm heading on to the main section tomorrow.